build: fix up legacy tooltip demo
Fixes the file and class names in the legacy tooltip demo and moves the
examples into the dev app.
