--- src/components-examples/material/legacy-tooltip/BUILD.bazel
@@ -1,58 +1,0 @@
-load("//tools:defaults.bzl", "ng_module", "ng_test_library", "ng_web_test_suite")
-
-package(default_visibility = ["//visibility:public"])
-
-ng_module(
-    name = "legacy-tooltip",
-    srcs = glob(
-        ["**/*.ts"],
-        exclude = ["**/*.spec.ts"],
-    ),
-    assets = glob([
-        "**/*.html",
-        "**/*.css",
-    ]),
-    deps = [
-        "//src/cdk/scrolling",
-        "//src/cdk/testing",
-        "//src/cdk/testing/testbed",
-        "//src/material/legacy-button",
-        "//src/material/legacy-checkbox",
-        "//src/material/legacy-input",
-        "//src/material/legacy-select",
-        "//src/material/legacy-tooltip",
-        "//src/material/legacy-tooltip/testing",
-        "@npm//@angular/forms",
-        "@npm//@angular/platform-browser",
-        "@npm//@angular/platform-browser-dynamic",
-        "@npm//@types/jasmine",
-    ],
-)
-
-filegroup(
-    name = "source-files",
-    srcs = glob([
-        "**/*.html",
-        "**/*.css",
-        "**/*.ts",
-    ]),
-)
-
-ng_test_library(
-    name = "unit_tests_lib",
-    srcs = glob(["**/*.spec.ts"]),
-    deps = [
-        ":legacy-tooltip",
-        "//src/cdk/testing",
-        "//src/cdk/testing/testbed",
-        "//src/material/legacy-tooltip",
-        "//src/material/legacy-tooltip/testing",
-        "@npm//@angular/platform-browser",
-        "@npm//@angular/platform-browser-dynamic",
-    ],
-)
-
-ng_web_test_suite(
-    name = "unit_tests",
-    deps = [":unit_tests_lib"],
-)
