--- src/components-examples/material/legacy-tooltip/legacy-tooltip-harness/legacy-tooltip-harness-example.html
@@ -1,2 +1,0 @@
-<button [matTooltip]="message" id="one">Trigger 1</button>
-<button matTooltip="Static message" id="two">Trigger 2</button>
