--- src/components-examples/material/legacy-tooltip/legacy-tooltip-harness/legacy-tooltip-harness-example.ts
@@ -1,12 +1,0 @@
-import {Component} from '@angular/core';
-
-/**
- * @title Testing with MatTooltipHarness
- */
-@Component({
-  selector: 'legacy-tooltip-harness-example',
-  templateUrl: 'legacy-tooltip-harness-example.html',
-})
-export class LegacyTooltipHarnessExample {
-  message = 'Tooltip message';
-}
