--- src/components-examples/material/legacy-tooltip/legacy-tooltip-manual/legacy-tooltip-manual-example.ts
@@ -1,11 +1,0 @@
-import {Component} from '@angular/core';
-
-/**
- * @title Tooltip that can be manually shown/hidden.
- */
-@Component({
-  selector: 'legacy-tooltip-manual-example',
-  templateUrl: 'legacy-tooltip-manual-example.html',
-  styleUrls: ['legacy-tooltip-manual-example.css'],
-})
-export class LegacyTooltipManualExample {}
