+++ src/dev-app/legacy-tooltip/examples/legacy-tooltip-auto-hide/legacy-tooltip-auto-hide-example.css
