+++ src/dev-app/legacy-tooltip/examples/legacy-tooltip-custom-class/legacy-tooltip-custom-class-example.ts
@@ -1,3 +1,11 @@
+/**
+ * @license
+ * Copyright Google LLC All Rights Reserved.
+ *
+ * Use of this source code is governed by an MIT-style license that can be
+ * found in the LICENSE file at https://angular.io/license
+ */
+
 import {Component, ViewEncapsulation} from '@angular/core';
 
 /**
