+++ src/dev-app/legacy-tooltip/examples/legacy-tooltip-manual/legacy-tooltip-manual-example.ts
@@ -5,13 +5,15 @@
  * Use of this source code is governed by an MIT-style license that can be
  * found in the LICENSE file at https://angular.io/license
  */
-import {Component} from '@angular/core';
-import {LegacyTooltipExamplesModule} from '@angular/components-examples/material/legacy-tooltip';
 
+import {Component} from '@angular/core';
+
+/**
+ * @title Tooltip that can be manually shown/hidden.
+ */
 @Component({
-  selector: 'tooltip-demo',
-  templateUrl: 'tooltip-demo.html',
-  standalone: true,
-  imports: [LegacyTooltipExamplesModule],
+  selector: 'legacy-tooltip-manual-example',
+  templateUrl: 'legacy-tooltip-manual-example.html',
+  styleUrls: ['legacy-tooltip-manual-example.css'],
 })
-export class TooltipDemo {}
+export class LegacyTooltipManualExample {}
