+++ src/dev-app/legacy-tooltip/examples/legacy-tooltip-position-at-origin/legacy-tooltip-position-at-origin-example.css
