+++ src/material/legacy-tooltip/BUILD.bazel
@@ -1,6 +1,5 @@
 load(
     "//tools:defaults.bzl",
-    "markdown_to_html",
     "ng_module",
     "ng_test_library",
     "ng_web_test_suite",
@@ -72,11 +71,6 @@ ng_test_library(
 ng_web_test_suite(
     name = "unit_tests",
     deps = [":unit_test_sources"],
-)
-
-markdown_to_html(
-    name = "overview",
-    srcs = [":tooltip.md"],
 )
 
 filegroup(
